Fixed update_source.sh script
Objenesis has moved from googlecode to github and the structure
has changed slightly; see previous commit. This change updates
the script to match. It also will update the README.version file
to ensure that it is up to date.

The tck-android/AndroidManifest.xml has moved to
tck-android/src/main/AndroidManifest.xml this change updates the
script to move it back in order for it to be found when
building ObjenesisTck.

Bug: 32912773
Test: Ran with ./update_source.sh 2.5 - see following commit
Change-Id: Id4871f31d85c3be5ae13240de1012af2cc0cf53e