Restructure files to match upstream objenesis
Makes the file structure match upstream objenesis so that it is
simpler to update in future. Moving the files as a separate
step to upgrading makes it easier for git to track renames and
simplifies future diffs and reviews.

Bug: 32912773
Test: make checkbuild

(cherry picked from commit d3bfa334632023b5401df247e60cb43844cdaa72)

Change-Id: I9a1b61c6967f3a3d1f618ce5c748281310e7b9f1